update: move init_key into src/suit_cose.c
