+++ Makefile.encode
@@ -5,8 +5,8 @@
 #
 
 CFLAGS = -Wall -g
-LDFLAGS = -lt_cose -lqcbor -lm
-INC = -I ./inc -I ./examples/inc
+LDFLAGS = $(CMD_LD) -lt_cose -lqcbor -lm
+INC = $(CMD_INC) -I ./inc -I ./examples/inc
 TARGET = ./bin/suit_manifest_encode
 SRCS = \
 	examples/suit_manifest_encode_main.c \
