+++ Makefile.parser
@@ -5,8 +5,8 @@
 #
 
 CFLAGS = -Wall -g
-LDFLAGS = -lt_cose -lqcbor -lm
-INC = -I ./inc -I ./examples/inc
+LDFLAGS = $(CMD_LD) -lt_cose -lqcbor -lm
+INC = $(CMD_INC) -I ./inc -I ./examples/inc
 TARGET = ./bin/suit_manifest_parser
 SRCS = \
 	examples/suit_manifest_parser_main.c \
